Refactor: RepositoryCustom 사용법 변경
- 테스트를 위해 상속받도록 변경
